requirements.txt
# Core Dependencies
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web & API
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Database Support
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pylint>=2.17.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Performance Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Machine Learning (for advanced insights)
scikit-learn>=1.3.0
statsmodels>=0.14.0

# API Development
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
tenacity>=8.2.0  # For retry mechanisms

# UI Framework
streamlit>=1.28.0
streamlit-aggrid>=0.3.4
